+++ lib/api/src/sys/store.rs
@@ -1,8 +1,11 @@
 use crate::sys::tunables::BaseTunables;
 use std::fmt;
 use std::sync::{Arc, RwLock};
+#[cfg(any(feature = "compiler", feature = "default-compiler"))]
 use wasmer_compiler::CompilerConfig;
-use wasmer_compiler::{Engine, Tunables, Universal};
+#[cfg(any(feature = "compiler", feature = "default-compiler"))]
+use wasmer_compiler::Universal;
+use wasmer_compiler::{Engine, Tunables};
 use wasmer_vm::{init_traps, TrapHandler, TrapHandlerFn};
 
 /// The store represents all global state that can be manipulated by
@@ -23,6 +26,7 @@ pub struct Store {
 }
 
 impl Store {
+    #[cfg(any(feature = "compiler", feature = "default-compiler"))]
     /// Creates a new `Store` with a specific [`CompilerConfig`].
     pub fn new(compiler_config: Box<dyn CompilerConfig>) -> Self {
         let engine = Universal::new(compiler_config).engine();